Перевод клиента 4.9.1 на схему API layer 228
Сетевой слой переписан целиком: разбор современной схемы сгенерирован
заново (legacy/TL/ModernTL_*), а к классам приложения 2018 года проложен
мост, который переносит поля по именам — код выше уровня TL не тронут.

Основное:
- запросы и отправляемые типы переведены на конструкторы layer 228;
- идентификаторы пиров расширены до 64 бит по всему приложению: у
  аккаунтов после 2021 года они не помещались в int32, из-за чего в
  списке чатов оставались пустые строки, а по нажатию на автора
  сообщения открывался профиль группы;
- голосовые звонки: phone.requestCall уходил без обязательного поля
  protocol, а сам phoneCallProtocol сериализовался конструктором старого
  слоя — звонок обрывался на "call failed";
- стикеры: webp снова распознаются как стикеры, анимированные .tgs
  рисуются через Lottie;
- Siri и iCloud не вызываются без соответствующего разрешения в подписи:
  системный assert внутри dispatch_once валил приложение при запуске и
  перехвату не поддаётся.

Исходники MtProtoKit и SSignalKit вшиты в репозиторий: они были
подмодулями, а правки сетевого слоя лежали в них незакоммиченными.
Добавлены LICENSE (GPL-2.0), README со сборкой и config.example.h —
ключи приложения читаются из ../../config.h вне репозитория.

+++ LegacyDatabase/TGChannelChatModel.h
@@ -9,6 +9,6 @@
 @property (nonatomic, readonly) bool isGroup;
 @property (nonatomic, readonly) int64_t accessHash;
 
-- (instancetype)initWithChannelId:(int32_t)channelId title:(NSString *)title avatarLocation:(TGFileLocation *)avatarLocation isGroup:(bool)isGroup accessHash:(int64_t)accessHash;
+- (instancetype)initWithChannelId:(int64_t)channelId title:(NSString *)title avatarLocation:(TGFileLocation *)avatarLocation isGroup:(bool)isGroup accessHash:(int64_t)accessHash;
 
 @end
