+++ LegacyDatabase/TGLegacyDatabase.h
@@ -13,7 +13,7 @@
 - (NSArray<TGLegacyUser *> *)topUsers;
 - (NSDictionary<NSNumber *, NSNumber *> *)unreadCountsForUsers:(NSArray<TGLegacyUser *> *)users;
 
-- (TGLegacyUser *)userWithIdSync:(int32_t)userId;
+- (TGLegacyUser *)userWithIdSync:(int64_t)userId;
 - (TGChatModel *)conversationWithIdSync:(int64_t)conversationId;
 
 - (NSData *)customPropertySync:(NSString *)name;
