+++ LegacyDatabase/TGLegacyDatabase.m
@@ -15,7 +15,7 @@ static inline bool TGPeerIdIsChannel(int64_t peerId) {
     return peerId <= ((int64_t)INT32_MIN) * 2 && peerId > ((int64_t)INT32_MIN) * 3;
 }
 
-static inline int64_t TGPeerIdFromChannelId(int32_t channelId) {
+static inline int64_t TGPeerIdFromChannelId(int64_t channelId) {
     return ((int64_t)INT32_MIN) * 2 - ((int64_t)channelId);
 }
 
@@ -220,7 +220,7 @@ int32_t murMurHash32(NSString *string)
             
             FMResultSet *result = [_database executeQuery:[NSString stringWithFormat:@"SELECT uid, first_name, last_name, local_first_name, local_last_name, phone_number, access_hash FROM users_v29 WHERE uid IN (SELECT uid FROM contacts_v29)"]];
             while ([result next]) {
-                int32_t uid = [result intForColumnIndex:0];
+                int64_t uid = [result intForColumnIndex:0];
                 NSString *firstName = [result stringForColumnIndex:1];
                 NSString *lastName = [result stringForColumnIndex:2];
                 NSString *phonebookFirstName = [result stringForColumnIndex:3];
@@ -309,7 +309,7 @@ int32_t murMurHash32(NSString *string)
         
         FMResultSet *result = [_database executeQuery:[NSString stringWithFormat:@"SELECT uid, first_name, last_name, local_first_name, local_last_name, phone_number, access_hash FROM users_v29 WHERE uid IN (SELECT uid FROM contacts_v29)"]];
         while ([result next]) {
-            int32_t uid = [result intForColumnIndex:0];
+            int64_t uid = [result intForColumnIndex:0];
             NSString *firstName = [result stringForColumnIndex:1];
             NSString *lastName = [result stringForColumnIndex:2];
             NSString *phonebookFirstName = [result stringForColumnIndex:3];
@@ -344,7 +344,7 @@ int32_t murMurHash32(NSString *string)
     
     FMResultSet *result = [_database executeQuery:[NSString stringWithFormat:@"SELECT u.uid, u.first_name, u.last_name, u.access_hash, u.photo_small FROM users_v29 u JOIN peer_rating_29 p ON u.uid = p.peer_id WHERE p.category = 1 ORDER BY p.rating DESC LIMIT 9"]];
     while ([result next]) {
-        int32_t uid = [result intForColumnIndex:0];
+        int64_t uid = [result intForColumnIndex:0];
         NSString *firstName = [result stringForColumnIndex:1];
         NSString *lastName = [result stringForColumnIndex:2];
         int64_t accessHash = [result intForColumnIndex:3];
@@ -383,7 +383,7 @@ int32_t murMurHash32(NSString *string)
     return counts;
 }
 
-- (TGLegacyUser *)userWithIdSync:(int32_t)userId {
+- (TGLegacyUser *)userWithIdSync:(int64_t)userId {
     
     __block TGLegacyUser *resultUser = nil;
     [_queue dispatchSync:^{
@@ -393,7 +393,7 @@ int32_t murMurHash32(NSString *string)
         FMResultSet *result = [_database executeQuery:[NSString stringWithFormat:@"SELECT uid, first_name, last_name, access_hash, photo_small, phone_number FROM %@ WHERE uid=?", tableName], @(userId)];
         if ([result next])
         {
-            int32_t uid = [result intForColumnIndex:0];
+            int64_t uid = [result intForColumnIndex:0];
             NSString *firstName = [result stringForColumnIndex:1];
             NSString *lastName = [result stringForColumnIndex:2];
             int64_t accessHash = [result intForColumnIndex:3];
