+++ LegacyDatabase/TGLegacyUser.m
@@ -2,7 +2,7 @@
 
 @implementation TGLegacyUser
 
-- (instancetype)initWithUserId:(int32_t)userId accessHash:(int64_t)accessHash firstName:(NSString *)firstName lastName:(NSString *)lastName phoneNumber:(NSString *)phoneNumber photoSmall:(NSString *)photoSmall {
+- (instancetype)initWithUserId:(int64_t)userId accessHash:(int64_t)accessHash firstName:(NSString *)firstName lastName:(NSString *)lastName phoneNumber:(NSString *)phoneNumber photoSmall:(NSString *)photoSmall {
     self = [super init];
     if (self != nil) {
         _userId = userId;
