+++ LegacyDatabase/TGPrivateChatModel.h
@@ -2,6 +2,6 @@
 
 @interface TGPrivateChatModel : TGChatModel
 
-- (instancetype)initWithUserId:(int32_t)userId;
+- (instancetype)initWithUserId:(int64_t)userId;
 
 @end
