+++ LegacyDatabase/TGUserModel.h
@@ -6,7 +6,7 @@
 
 @interface TGUserModel : NSObject
 
-@property (nonatomic, readonly) int32_t userId;
+@property (nonatomic, readonly) int64_t userId;
 @property (nonatomic, readonly) int64_t accessHash;
 
 @property (nonatomic, strong, readonly) NSString *firstName;
@@ -14,7 +14,7 @@
 
 @property (nonatomic, strong, readonly) TGFileLocation *avatarLocation;
 
-- (instancetype)initWithUserId:(int32_t)userId accessHash:(int64_t)accessHash firstName:(NSString *)firstName lastName:(NSString *)lastName avatarLocation:(TGFileLocation *)avatarLocation;
+- (instancetype)initWithUserId:(int64_t)userId accessHash:(int64_t)accessHash firstName:(NSString *)firstName lastName:(NSString *)lastName avatarLocation:(TGFileLocation *)avatarLocation;
 
 - (NSString *)displayName;
