+++ Telegraph/TGBotContextResults.h
@@ -11,7 +11,7 @@
 
 @interface TGBotContextResults : NSObject <NSCoding>
 
-@property (nonatomic, readonly) int32_t userId;
+@property (nonatomic, readonly) int64_t userId;
 @property (nonatomic, readonly) int64_t peerId;
 @property (nonatomic, readonly) int64_t accessHash;
 
@@ -21,6 +21,6 @@
 @property (nonatomic, strong, readonly) NSArray *results;
 @property (nonatomic, strong, readonly) TGBotContextResultsSwitchPm *switchPm;
 
-- (instancetype)initWithUserId:(int32_t)userId peerId:(int64_t)peerId accessHash:(int64_t)accessHash isMedia:(bool)isMedia query:(NSString *)query nextOffset:(NSString *)nextOffset results:(NSArray *)results switchPm:(TGBotContextResultsSwitchPm *)switchPm;
+- (instancetype)initWithUserId:(int64_t)userId peerId:(int64_t)peerId accessHash:(int64_t)accessHash isMedia:(bool)isMedia query:(NSString *)query nextOffset:(NSString *)nextOffset results:(NSArray *)results switchPm:(TGBotContextResultsSwitchPm *)switchPm;
 
 @end
